use bigint rather than integer, raise cap to 100m

+++ db/migrate/20241022214312_add_untrusted_favourites_count_and_untrusted_reblogs_count_to_status_stat.rb
@@ -2,7 +2,7 @@
 
 class AddUntrustedFavouritesCountAndUntrustedReblogsCountToStatusStat < ActiveRecord::Migration[7.1]
   def change
-    add_column :status_stats, :untrusted_favourites_count, :integer, null: true
-    add_column :status_stats, :untrusted_reblogs_count, :integer, null: true
+    add_column :status_stats, :untrusted_favourites_count, :bigint, null: true
+    add_column :status_stats, :untrusted_reblogs_count, :bigint, null: true
   end
 end
